+++ selene/__init__.py
@@ -22,12 +22,12 @@
 
 
 from selene.core.entity import (
-    Browser as _custom_browser,
-    Config as _custom_config_for_custom_browser,
+    Browser as _CustomBrowser,
+    Config as _CustomConfigForCustomBrowser,
 )
 
-Config = _custom_config_for_custom_browser
-Browser = _custom_browser
+Config = _CustomConfigForCustomBrowser
+Browser = _CustomBrowser
 """
 Given::
 
@@ -56,7 +56,7 @@ AND::
 """
 
 
-from selene.support import by as _by_style_selectors
+from selene.support import by as _by_style_selectors  # noqa
 
 by = _by_style_selectors
 """
@@ -81,7 +81,7 @@ AND::
 """
 
 
-from selene.support.conditions import be as _be_style_conditions
+from selene.support.conditions import be as _be_style_conditions  # noqa
 
 be = _be_style_conditions
 """
@@ -93,7 +93,7 @@ AND (in case we need to filter collection of items by some condition like visibi
 """
 
 
-from selene.support.conditions import have as _have_style_conditions
+from selene.support.conditions import have as _have_style_conditions  # noqa
 
 have = _have_style_conditions
 """
@@ -116,7 +116,7 @@ FINALLY (if not registered "atexit" before)::
 # Advanced Helpers #  # todo: think on not adding them here...
 ####################
 
-from selene.core import command as _advanced_commands
+from selene.core import command as _advanced_commands  # noqa
 
 command = _advanced_commands
 """
@@ -128,7 +128,7 @@ e.g. for workaround something through js::
     browser.element('#not-in-view').perform(command.js.scroll_into_view)
 """
 
-from selene.core import query as _advanced_queries
+from selene.core import query as _advanced_queries  # noqa
 
 query = _advanced_queries
 # its = _advanced_queries  # todo: do we really need it too? for better readability: .get(its.text)
@@ -218,6 +218,6 @@ Or, by using non-waiting versions, if "you are in a rush:)"::
 # """
 # Just types...
 # """
-# todo: add here some type imports like Element, Collection, etc.
+from selene.core.entity import Element, Collection  # noqa
 
 __version__ = '2.0.0b11'